docker/requirements.txt
boto3==1.34.0
Flask-SQLAlchemy==3.0.5
Flask==2.3.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23